# lib/api-zod/src/generated/types/generatorRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Generator Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratorRecord {
  id: number;
  tDate: string;
  generatorId: string;
  status: string;
  /** @nullable */
  rating?: string | null;
  /** @nullable */
  hours?: number | null;
  /** @nullable */
  remarks?: string | null;
  createdAt: string;
  updatedAt: string;
}