# lib/api-zod/src/generated/types/generatorStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Generator Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratorStatsByStatusItem } from './generatorStatsByStatusItem';

export interface GeneratorStats {
  total: number;
  byStatus: GeneratorStatsByStatusItem[];
  /** @nullable */
  avgHours?: number | null;
  recentCount: number;
}